Add a notification if an svn_fs_pack call is a no-op.  Use that to warn users if they try to pack a non-sharded FSFS repo.
Extending the svn_fs_pack_notify_action_t enum seems to be the only way
to implement this without bumping the svn_fs_pack API:  We can't use the
warning function because the svn_fs_t instance is created temporarily
inside the API function such that the user can't set warning function.
No-op packs are not an error condition either, therefore returning an
svn_error_t would be inappropriate.

To get the most milage out of this feature, we also notify if there is
no complete shard that can be packed.  'svnadmin pack' does not report
that condition to minimize UI changes.  Other API users might be interested
in that information, though.

* subversion/include/svn_fs.h
  (svn_fs_pack_notify_action_t): Add notification for no-op "action".

* subversion/include/svn_repos.h
  (svn_repos_notify_action_t): Same in outer API layer.

* subversion/libsvn_fs_fs/pack.c
  (pack_body): Notify if there was no shard to be packed.
  (svn_fs_fs__pack): Same.  Also warn if the repo is non-sharded.

* subversion/libsvn_repos/fs-wrap.c
  (pack_notify_func): Translate the new notification as well.

* subversion/svnadmin/svnadmin.c
  (repos_notify_handler): Print a warning if the repo is non-sharded.

* subversion/tests/cmdline/svnadmin_tests.py
  (fsfs_pack_non_sharded): New test case.
  (test_list): Register the new test.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1688270 13f79535-47bb-0310-9956-ffa450edef68