It is possible to run the svn-mergeinfo-normalizer tool on a working copy that refers to a path that does not exist in the repository (anymore).
Instead of erroring out, we should simply try its parent paths up to the
repository root until we find one that does exist.  Then run the log
request on that one.  We will still get error if repo does not exist.  The
"extra" log info is a mere inefficiency that does not affect functionality.

* tools/client-side/svn-mergeinfo-normalizer/logic.c
  (get_url): New utility function.
  (svn_min__run_normalize): Also open a session if we just need log access.
                            Call to new function to read the log from a
                            URL we know actually exists.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1696183 13f79535-47bb-0310-9956-ffa450edef68