Tune reorg strategy during the FSFS format7 packs such that it favors checkout-style tree walks now.
Since r1703237, following the log history no longer requires frequent
access to directory data but mainly relies on noderev predecessor chain.
Therefore, it is no longer necessary to tightly pack directories in a
separate part of the pack file.  With this patch, they are now placed
with the file contents and can be processed by a quasi-linear scan
instead of reading from two sections per pack.

* subversion/libsvn_fs_fs/pack.c
  (compare_dir_entries_format7): Adapt reporting strategy - process dirs
                                 at the same time as files now.
  (compare_is_dir): No longer needed.
  (sort_reps): No longer distinguish between file and dir reps but only
               paths and delta chains when determining reprentation order.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1705646 13f79535-47bb-0310-9956-ffa450edef68