Resolve issue #4597, and add a regression test for this problem.
The problem here is that our diff code doesn't handle full url, revision, peg
revision walking. The code assumes that the user either passes a single path
and upto 3 revisions -left, right, peg- or a path/url and a single revision.

The internal apis guessed which way they were called by looking at specific
argument combinations. In the 1.9 refactoring of diff to the common diff tree
processors the code was normalized to how we usually handle these revisions
in other code, fixing theoretical issues, but losing the notion that this
code is only used in a limited number of cases...

In the specific case of issue #4597 no peg revision was passed, but just a
(url/local path)+ revision... But because we still had an 'unspecified' peg
revision we applied the default rule, resolve revision in HEAD.

The patch resolves this problem by adding an additional argument telling that
we are in this specific case.

* subversion/libsvn_client/diff.c
  (diff_repos_wc): Update arguments to more closely document behavior.
  (do_diff): Add no_peg_revision argument. Update caller.
  (svn_client_diff6,
   svn_client_diff_peg6,
   svn_client_diff_summarize2,
   svn_client_diff_summarize_peg2): Update caller.

* subversion/tests/cmdline/diff_tests.py
  (diff_peg_resolve): New test.
  (test_list): Add diff_peg_resolve.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1707493 13f79535-47bb-0310-9956-ffa450edef68