Speed up uri_escape() even further by making the EOS test implicit.
Because all strings are NUL-terminated C strings, we can simply mark NUL
as "invalid" in any table we pass to uri_escape; it will never be used
within a string.  Then, we only need to scan up to the first "invalid"
char and then check it for NUL.

This can significantly improve the throughput on newer processors
(Haswell and up) but others will see some benefit, too.

* subversion/libsvn_subr/path.c
  (uri_escape): Verify that the table meets our assumption about NULs
                and then scan the PATH without calculating its length
                first.
  (iri_escape_chars,
   uri_autoescape_chars): Mark NUL as "invalid".

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1710103 13f79535-47bb-0310-9956-ffa450edef68