Minor refactoring of FSFS' L1 dag node cache access.
Split lookup and insertion into two separate functions.  That comes at
the cost of calculating the hash twice but allows us to delay the cache
auto-clear call until we actually have an L2 item that we can promote
to L1.  The redcution in cache misses will often compensate for the hash
calculation overhead.

* subversion/libsvn_fs_fs/tree.c
  (hash_func): Move out of cache_lookup.
  (cache_lookup): Don't clear the entry when we found a mismatch and
                  return the node - if found - instead of the whole bucket.
  (cache_insert): New function.
  (dag_node_cache_get): Update and slightly simplify the caller.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1710368 13f79535-47bb-0310-9956-ffa450edef68