In FSX, make revprop modifications use the batch fsync infrastructure.
This does not make a difference in most cases but is part of the fsync
handling unification effort.  We also use well-defined file names for
the temp files involved as they guarantee volume-local renames and are
easier to use with svn_fs_x__batch_fsync_open_file.  They are safe
because we know that we have exclusive write access to the repo.

* subversion/libsvn_fs_x/revprops.c
  (write_non_packed_revprop): Implicitly schedule an fsync through the
                              provided BATCH and append ".tmp" to form
                              temp file names.
  (switch_to_new_revprop): Implement the fsync-rename-fsync cycle based
                           on an fsync BATCH now.
  (repack_revprops): All revprop files are fsync'ed implicitly through
                     some batch now.
  (repack_file_open): Implicitly schedule an fsync through the BATCH.
  (write_packed_revprop): Same. Pass BATCH into sub-routines.
  (svn_fs_x__set_revision_proplist): Create the fsync batch locally.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1711369 13f79535-47bb-0310-9956-ffa450edef68