[Reverted in 1717864]
The svn_fs_contents_different implementation is broken in 1.9.x under the
circumstances described in issue #4554.  Trunk works because there we
correct expanded-size values when reading them from disk.

This patch removes the special case code that triggers the problem in 1.9.x
so we can backport it directly as a fix.  On trunk, it results in a minor
simplification with virtually zero impact on performance.

The problem is FSFS-specific, meaning BDB and FSX don't need to be fixed.

Found by: kotkov

* subversion/libsvn_fs_fs/fs_fs.c
  (svn_fs_fs__file_text_rep_equal): Remove the rather pointless length==0
                                    special case. They are rare and will
                                    compare just fine like any other contents.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1712927 13f79535-47bb-0310-9956-ffa450edef68