Teach 'svnmover' to support a mixed-revision WC.
The WC storage structure now supports a different base revision for each
element. The 'info-wc' command now shows minimum and maximum base revisions
(if they differ). There is not yet a way to see the base revision of a
specific element.

Commit will now produce a mixed-rev WC: it no longer updates the whole WC.

In order to fully identify a base element with an older revision number, we
will need a way to trace the branch back through history, following any
changes in the branch id.

Bug: The committed revision doesn't get attached to the WC's 'repos' object
(only to the server's 'repos' object) so we then can't access it: e.g.
'update .@HEAD' fails with 'No such revision'.

* tools/dev/svnmover/svnmover.c
  (svnmover_wc_set_base_rev,
   svnmover_wc_get_base_rev,
   svnmover_wc_set_base_revs,
   svnmover_wc_get_base_revs): New.
  (wc_checkout): Initialize the base revision numbers.
  (update_wc_eids,
   update_wc_base): New.
  (wc_commit): Update the WC base to reflect the commit.
  (find_el_rev_by_rrpath_rev): Allow looking up an element in revision
    'BASE'.
  (subtree_diff_r,
   branch_diff_r): Stop reporting a (single) revision number for each side
    of the diff.
  (commit): Tweak a comment.
  (do_commit): Don't update the whole WC after commit.
  (arg_t): Remove the 'revnum' field.
  (execute): Expand the 'info-wc' display to report the minimum and maximum
    base revision numbers. Track the other changes. Add a comment.

* tools/dev/svnmover/svnmover.h
  (svnmover_wc_t): Add storage for the base revision numbers.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1716710 13f79535-47bb-0310-9956-ffa450edef68