In 'svnmover', improve how we diff and merge subtrees.
When merging a subtree, notice all changes to each element that is in any of
the three subtrees. In other words, for an element that is present in only
the source or only the target subtree, don't ignore any changes that may
have happened to that element on the other side of the merge, even though it
is outside the given subtrees on that side. Add a regression test for this.

When merging an element that is the branch root in one or two of the three
(YCA, source, target) inputs, but is a non-root element in others, ignore
that element's parentage. Otherwise, if we try to merge that difference in
parentage, we would mess up the tree structure of the target branch.
svnmover_tests.py 16 "merge from subbranch to subtree" covers this case, and
would fail if only the "notice all changes" change above were committed.

* tools/dev/svnmover/merge3.c
  (is_branch_root_element): New, copied from svnmover.c.
  (branch_merge_subtree_r): Diff the elements in the union of all three
    subtrees, not just those in each pair of subtrees.

* tools/dev/svnmover/svnmover.h
  (svnmover_element_differences): Take the set of elements to diff as an
    input. Allow other inputs to be null.

* tools/dev/svnmover/svnmover.c
  (svnmover_element_differences): Take the set of elements to diff as an
    input. Allow other inputs to be null.
  (txn_is_changed): Update caller, passing null.
  (branch_elements_replay): Rename from 'subtree_replay'. Take the set of
    elements to diff as an input. Simplify.
  (svn_branch__replay,
   update_wc_base_r,
   subtree_diff): Update callers, passing null.
  (show_subtree_diff): Update the doc string.

* subversion/tests/cmdline/svnmover_tests.py
  (merge_move_into_subtree): New test.
  (test_list): Run it.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1717778 13f79535-47bb-0310-9956-ffa450edef68