Following up on r1717874, make all ra layers consistently report inherited properties in the documented way.
This makes ra_svn and ra_serf against new servers report the properties in
the documented way and the way ra_local already reported these.

Which form ra_serf reported before this patch depended on the server
capabilities... If it could use the optimized request it reported like
ra_svn, if not like ra_local.

* subversion/libsvn_ra_serf/inherited_props.c
  (iprops_closed): Stop adding the repository url in output arguments in
    unexpected ways.

* subversion/libsvn_ra_svn/client.c
  (parse_iproplist): Stop adding the repository url in output arguments
    when it is documented that we shouldn't do this.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1717875 13f79535-47bb-0310-9956-ffa450edef68